+++ solutions/038.cpp
@@ -18,7 +18,7 @@ std::string concat(const std::array<int, 10>& arr, size_t n) {
 
 bool concatProduct(const std::string& s) {
   for (size_t i = 1; i < s.size(); i++) {
-    int base = std::stod(s.substr(0, i));
+    int base = std::stoi(s.substr(0, i));
     size_t n = s.size() / i;
     std::array<int, 10> arr;
     for (size_t j = 0; j < n; j++) {
@@ -36,5 +36,5 @@ bool concatProduct(const std::string& s) {
 int p38() {
   auto is = permutations("123456789");
   auto it = std::remove_if(is.begin(), is.end(), std::not_fn(concatProduct));
-  return std::stod(*std::max_element(is.begin(), it));
+  return std::stoi(*std::max_element(is.begin(), it));
 }
